# docs/swagger.yaml
definitions:
  entity.Authentication:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  entity.PostUser:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
    type: object
  entity.User:
    properties:
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
  title: Life Achieve
  version: "1.0"
paths:
  /signin:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authentication object
        in: body
        name: authentication
        required: true
        schema:
          $ref: '#/definitions/entity.Authentication'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
      summary: Get jwt token
      tags:
      - auth
  /signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.PostUser'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
      summary: Post user
      tags:
      - auth
  /users/{id}:
    get:
      parameters:
      - description: JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.User'
            type: array
        "400":
          description: Bad Request
      summary: Get user
      tags:
      - users
    patch:
      consumes:
      - application/json
      parameters:
      - description: JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
      summary: Patch user
      tags:
      - users
swagger: "2.0"